- *ADDED:* You can now use Azure Container Apps as a compute platform for leveraging database builds with the new `sbm containerapps` commands. See the [Azure Container Apps documentation](containerapps.md) for background, information and how-to examples - *ADDED:* New options for container image tags and container registries for all container options (Azure Container Apps, Kubernetes, Azure Container Instance) - *UPDATED:* Updates to deployment scripts and unit tests - *ADDED:* New command `sbm utility eventhub` to scan EventHub for job event counts - *UPDATED:* Reorganized template scripts by service - *UPDATED:* Optimized EventHub client to checkpoint by job time so events are not missed when running multiple concurrent jobs

+++ scripts/templates/Batch/create_batch_settingsfiles.ps1
@@ -47,13 +47,15 @@ $settingsJsonLinuxQueue = Join-Path $path "settingsfile-linux-queue.json"
 $settingsJsonWindowsQueueKv = Join-Path $path "settingsfile-windows-queue-keyvault.json"
 $settingsJsonLinuxQueueKv = Join-Path $path "settingsfile-linux-queue-keyvault.json"
 
-
-$AESKey = New-Object Byte[] 32
-[Security.Cryptography.RNGCryptoServiceProvider]::Create().GetBytes($AESKey)
-$settingsFileKey = [System.Convert]::ToBase64String($AESKey);
-
 $keyFile = Join-Path $path "settingsfilekey.txt"
-$settingsFileKey |  Set-Content -Path $keyFile
+if($false -eq (Test-Path $keyFile))
+{
+    $AESKey = New-Object Byte[] 32
+    [Security.Cryptography.RNGCryptoServiceProvider]::Create().GetBytes($AESKey)
+    $settingsFileKey = [System.Convert]::ToBase64String($AESKey);
+    $settingsFileKey |  Set-Content -Path $keyFile
+}
+
 
 $winParams = @("--batchpoolname","SqlBuildManagerPoolWindows")
 $winParams += ("-os", "Windows")
